requirements.txt
flask
flask-restful
flask-cors
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
requests
bcrypt
